+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableBooleanArray.kt
@@ -209,6 +209,11 @@ public value class ImmutableBooleanArray @PublishedApi internal constructor(
     }
 
     /**
+     * See [BooleanArray.random]
+     */
+    public inline fun random(): Boolean = values.random()
+
+    /**
      * See [BooleanArray.iterator]
      */
     public inline operator fun iterator(): Iterator<Boolean> = values.iterator()